Split scaffold test environment helper boundaries
